mlflow_code/requirements.txt
mlflow==2.13.0
boto3
pandas
scikit-learn
python-dotenv
SQLAlchemy
psycopg2-binary